# packages/contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: 전공한시간 업무 API
  version: "0.2.2-adhoc-curriculum-draft"
  description: >
    S04-T01 백엔드 엔드포인트 및 DTO 반영본. S02 베이스라인에 누락되었던
    sampleDescription, /tutors/{id}, /tutor/courses/{id} PATCH, availability,
    /admin/learning-requests/{id}/assist, /subjects, /life-zones를 추가했다.
    0.2.2: 에드혹 추가(사용자 요청, docs/handoffs/ADHOC-01-ai-curriculum-draft.md) —
    Course.curriculumSource, POST /tutor/courses/curriculum/draft. S04 GATE 이후
    별도 A0 통합 리뷰 없이 추가된 변경이므로 다음 정식 게이트에서 재검토가 필요하다.
servers:
  - url: /api
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  responses:
...
    Forbidden:
      description: 권한 없음(참여자·소유자·업무 권한 불일치)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - STALE_PROPOSAL
            - SCHEDULE_CONFLICT
            - PAYMENT_AMOUNT_MISMATCH
            - PAYMENT_RECONCILIATION_REQUIRED
            - IDEMPOTENCY_KEY_CONFLICT
            - SELF_DEALING_NOT_ALLOWED
            - CONDITION_NOT_MET
        message:
          type: string
        requestId:
          type: string
    Money:
      type: integer
      description: KRW 정수(원)
...
    VerificationBadge:
      type: object
      properties:
        type:
          type: string
          enum: [school_enrollment, major, career, certification]
        status:
          type: string
          enum: [self_reported, pending, verified, rejected]
    TutorPublicProfile:
      type: object
      properties:
        id: { type: string }
        displayName: { type: string, description: "이름 일부 표기" }
        school: { type: string, nullable: true }
        major: { type: string, nullable: true }
        verificationBadges:
          type: array
          items: { $ref: '#/components/schemas/VerificationBadge' }
        teachingEvidence:
          type: object
          description: "인증 배지와 분리된 필드 — authorization.md §5"
          properties:
            sampleDescription: { type: string }
            expectedOutcomes: { type: array, items: { type: string } }
            reviewSummary: { type: object }
...
    LearningRequest:
      type: object
      required: [id, learnerId, goal, lifeZoneId, status]
      properties:
        id: { type: string }
        learnerId: { type: string }
        goal: { type: string }
        level: { type: string, nullable: true }
        deadline: { type: string, format: date-time, nullable: true }
        budgetRange:
          type: object
          properties:
            min: { $ref: '#/components/schemas/Money' }
            max: { $ref: '#/components/schemas/Money' }
        desiredWindows:
          type: array
          items:
            type: object
            properties:
              start: { type: string, format: date-time }
              end: { type: string, format: date-time }
        lifeZoneId: { type: string }
        acquisitionChannel: { type: string, enum: [organic, community_post, referral_manual, paid_ad, direct], description: "AR-03" }
        status: { type: string, enum: [open, matched, waitlisted, expired, withdrawn] }
...
    WaitlistEntry:
      type: object
      properties:
        id: { type: string }
        learningRequestId: { type: string }
        learnerId: { type: string }
        alternativeTimeAccepted: { type: boolean }
        notifyConsent: { type: boolean, default: true, description: "알림 동의 여부 — S04-T01" }
        status: { type: string, enum: [open, notified, matched, expired, withdrawn] }
...
paths:
...
  /tutors/{id}:
    get:
      summary: 교육자 공개 프로필 조회(S04-T01)
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TutorPublicProfile' }
        "404": { description: 존재하지 않는 교육자 }
...
  /tutor/profile:
    post:
      summary: 교육자 자기기재 프로필 등록/수정(CR-02)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                selfReportedSchool: { type: string }
                selfReportedMajor: { type: string }
                selfReportedCareer: { type: array, items: { type: object } }
      responses:
        "200": { description: "저장됨. 이 값만으로 공개 프로필의 verified 배지가 생성되지 않음(authorization.md §5)" }
...
  /admin/learning-requests:
    get:
      summary: 학습 요청 매칭 대기열(ops.matching)
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LearningRequest' }
  /admin/waitlist-entries:
    get:
      summary: 대기 신청 목록(ops.matching)
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WaitlistEntry' }
...
  /learning-requests:
    post:
      summary: 학습 요청 등록(학습자 본인)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LearningRequest' }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LearningRequest' }
...
  /admin/learning-requests/{id}/assist:
    post:
      summary: 운영자 수동 매칭 보조 처리(S04-T01)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                matchedTutorId: { type: string }
                unmatchedReason: { type: string }
                note: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LearningRequest' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { description: 학습 요청 없음 }
...